Move `unobjctify` to utils and use throughout.
Added casting to datetime for NSDate.
